4.6_the_project_step23/src/todo-backend/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
psycopg[binary]
prometheus-fastapi-instrumentator
nats-py
